Rename packetSerial.cpp/h to hal_packetserial.cpp/h - Avoid issues with case-incensitive OS
