+++ lib/platform/common/packetSerial/hal_packetserial.h
